requirements.txt
requests
databutton
streamlit
spacy_streamlit
mysql-connector-python